backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
yfinance==0.2.65
pandas==2.1.3
python-dotenv==1.0.0